Implemented and integrated fast drawing to generator_frame
